Add Maximize Happiness of Selected Children

+++ Maximize_Happiness_of_Selected_Children.cpp
@@ -1,11 +1,10 @@
 class Solution {
 public:
     long long maximumHappinessSum(vector<int>& happiness, int k) {
-        long long ans=0;
+        int ans=0;
         sort(happiness.begin(),happiness.end(),greater<int>());
         for(int i=0;i<k;i++)
             ans += max(happiness[i]-i, 0);
         return ans;
     }
 };
-// Title: Maximize Happiness of Selected Children
